[gens-qt4] VBackend/: Optimize VBackend updating.
Don't reupload the MD screen texture if the MD screen hasn't changed.

Don't redraw the GL widget if nothing's changed.

TODO: Test this on all platforms. I may have to remove the second
optimization if it breaks on some platforms. (It seems to work on
Linux/X11 using DRI2 and Gallium3D.)

TODO: Figure out why dragging a window off the top of the GL widget
causes everything to slow down. (It even happens in rss-glx, so that
might just be an X11 issue.)