[gens-qt4] VBackend/GLShaderManager.cpp: Split the shaders into individual classes.
The shaders inherit from a base class, GLShader. This makes it
easier to manage the different types of shaders, including the
current GL_ARB_fragment_program and (broken) GL_ATI_text_fragment_shader.

It will also make it possible to add support for GLSL, which I'll
attempt to use instead of GL_ATI_text_fragment_shader for the Power Mac.
(The Radeon 9000 driver apparently supports GL_ARB_shading_language_100,
but I don't know if it actually supports GLSL or if that's just an
extension for querying GLSL capabilities.)