[gens-qt4] IdleThread.cpp: Reimplemented the Intro Effect using threads.
It's a bit of a hack job, but it seems to work.

The "Crazy" Effect is now implemented and working.

GensWindow will create an IdleThread to run the intro effect when
emulation is stopped. A bunch of checks were added to prevent the
idle thread from causing issues when emulation is started again
and when the program is closed, but it still might have some problems.

If the intro effect style is set to "None", then the idle thread
won't be used at all.

NOTE: If the intro effect style is not set to "None", then the
MD screen won't be cleared when emulation is stopped. This is done
to simulate how the "Crazy" Effect worked in old Gens, i.e. it
decomposed the last image that was shown onscreen.

TODO: The intro effect appears to be started if a new ROM is opened
while a ROM is already running. This shouldn't happen.