[libgens] VdpRend_m5::T_Render_Line_ScrollA(): Fixed window display for many games.
andlabs noticed that many games weren't showing the window region properly.

Elemental Master: Window was scrolling with Scroll B. (Turns out that
both Scroll B and the Window were set to the same area in VRAM, but
that was only a symptom of the bug.)

Vapor Trail: Window shows garbage.

Grind Stormer: Window is blank.

Fire Shark: Window initially works, but then is overwritten by the
main scroll planes.

Turns out the problem was the X offset wasn't being added for
right-aligned windows, so it was reading the wrong data.

This patch fixes the problem for the aforementioned games.

(My previous test ROM, Gauntlet 4, didn't show any problems because
it only used Scroll B for the main game area, and Scroll A was
basically ignored.  Scroll A and B were set to 0xE000, while Window
was set to 0xD000.)