[gens-qt4] KeyHandlerQt::KeyValMToQtKey(): Don't perform Ctrl/Cmd switching.
Use the Ctrl modifier in the configuration file as if it were Cmd on
Mac OS X. This makes it easier to deal with Qt's automatic Ctrl/Cmd
key swapping.

actions/GensKeyConfig.cpp: Set the default key for "General Configuration"
on Mac OS X to be KEYM_CTRL | KEYV_COMMA, which will be remapped to Cmd+,
by Qt automatically.