[gens-qt4] Reworked the VBackend subsystem such that VBackend now inherits from QWidget.
GensQGLWidget inherits from VBackend, but does not inherit from QGLWidget.
Instead, GensQGLWidget contains a GensQGLWidgetPrivate, a private widget
that inherits QGLWidget. The end result is the same, but the code is
much cleaner.

Moved most of the slots from GensQGLWidget to VBackend.

Moved event handling from GensQGLWidget to GensQGLWidgetPrivate.

Clenaed up a ton of code in GLBackend and VBackend, including constification
of various variables.

Moved inline functions declared in class definitions to outside of the
class definition, but still within the header file.