[gens-qt4] GeneralConfigWindow: Initial "System" tab.
Currently, the System tab contains configuration for setting the current
region code (also available in the "System" menu in the GensWindow)
and for setting the region auto-detection order.

Note: None of this is functional at the moment.

ui/resources/oxygen/: Added arrow-down.png and arrow-up.png for all
sizes except 64x64.

GeneralConfigWindow.cpp: Added code for the up/down buttons.

TODO: Detect drag-and-drop of items within the QListWidget.