[gens-qt4] GensConfig.cpp: Added a property regionCodeOrder.
This property is a 16-bit value that indicates the priority preferences
for region code auto-detection. The most-significant nybble indicates
the highest-priority region, and each lesser-significant nyblle indicates
the region of lower priorities.

There are only 24 possible valid values for this value. Each nybble
must have a single bit set, and each possible value (1, 2, 4, and 8)
must be used exactly once. If setRegionCodeOrder() is called with an
invalid value, it is ignored. If the configuration file has an invalid
value, a default value is used. (currently 0x4812 for US, EU, JP, Asia)