[gens-qt4] Added menu synchronization for Pause.
This commit has a lot of changes. Among other things, instead of a
single slot for pause, it uses a slot stateChanged(), which functions
the same way as GensWindow's stateChanged() slot.

GensActions::actionTogglePause() is now GensActions::actionSetPause(bool).
The menu check state is specified as the parameter.

EmuManager::pauseRequest() is now EmuManager::pauseRequest(bool).
It uses the specified parameter instead of toggling the manual pause state.

GensMenuBar now takes an m_emuManager pointer in the constructor. This is
needed in order to receive signals from the Emulation Manager.

GensMenuBar: Instead of disconnecting and reconnecting the signal mapper,
use a wrapper function to retrieve the mapped(int) signal. This is needed
in order to obtain the menu check state from the menu bar, which is then
emitted in the triggered(int, bool) signal. The wrapper function won't
do anything if the menu bar is locked.

There's probably some other changes that aren't too important to be listed
here, so check the commitdiff for more information.