[gens-qt4] CtrlConfigWindow: Added Team Player and EA 4-Way Play ports.
The ports are only visible if the correct devices are set for
Port 1 and Port 2.

Changed all instances of "Teamplayer" to "Team Player".

Team Player and 4WP ports only support 3-button and 6-button controllers.
(4WP technically supports anything, but it's more difficult to add support
for anything other than the basic two controllers, and no games that support
4WP support other devices anyway.)

TODO: Add 4WP configuration support. (It requires both Port 1 and
Port 2 to be set to the appropriate 4WP devices.)