[gens-qt4] EmuManager::sysName(): Added disambiguating comments for localized system names.
The system names should only be translated. For example, "Mega Drive" should
be converted to the appropriate katakana for ja_JP. However, "Genesis" should
not be changed to "Mega Drive" for non-US regions, since this string is based
on the ROM region, not the system locale.

translations/*.ts: Updated translation files for EmuManager::sysName() changes.