[WTF] [libgens] SRam::writeWord(): Fixed word writes to SRAM.
We were writing both halves of the data word to the low byte, which ended
up breaking Puggsy's copy-protection system.

Puggsy writes a word to SRAM a few times to check if it's running on a copier.
If the written word is read back correctly, it assumes it is and prevents
progression past the third level.

TODO: Add a hack to disable SRAM for Puggsy.