[gens-qt4] Input/KeyHandlerQt.cpp: Handle numeric keypad keys correctly.
Numeric keypad keys should be converted to the appropriate KEYV_KP*
keys, regardless of whether or not Num Lock is enabled.

TODO: Verify Mac OS X-specific aspects, e.g. the "Clear" key (which
is mapped to 5 when Num Lock is off on Linux) and the '=' key.

TODO: Verify numeric keypad modifiers on actions that are present
in menus. (Alternatively, just don't allow numeric keypad keys
to be used for actions at all, and reserve them exclusively for
controller configuration.)