[gens-qt4] AboutWindow.cpp: Added translation metadata.
Scarred Sun reported some issues with translating a few strings to Japanese.
In particular, "Save directory" was confusing, since it could either be
a prompt for the user to save a selected directory or indicating a path
for a savestate directory. (In this case, it's the latter.)

Also, fixed QT ASCII cast issues in GetCodePageInfo(). I'm not sure why
this didn't break the Win32 build, since I disabled casting ASCII strings
to QString...