[libgens] credits.c::GensGS_credits[]: Added credits for translations.
en_GB: Overlord
ja_JP: Scarred Sun

These credits use a new format for subtitles. The subtitle is in the
credit_title field, and it starts with a '*'. This indicates that
the credit should be handled as a subtitle to the previous title,
and shouldn't be handled as an entirely new credit title.