[gens-qt4] Added changeEvent() functions to all windows to allow changing the language at runtime.
The changeEvent() functions call retranslateUi(). Some windows have additonal
text that needs to be translated in addition to the text specified in the
UI files, so that text is also retranslated in the changeEvent() function.

actions/GensMenuBar.cpp: Added a retranslate() function to retranslate
all of the menu items. This function is called by GensWindow::changeEvent().

gqt4_main.cpp: gqt4_qtTranslator and gqt4_gensTranslator are now available
outside of the gqt4_main.cpp file. The UI language can be changed by calling
the load() function on both translators.

TODO:
- Write a wrapper function to change the language on both translators.
- Add a UI to change the current language. (maybe in GeneralConfigWindow?)