[gens-qt4] CtrlConfigWindow: Updated the icon for "None".
The icon is a modified version of the previous icon. It now represents
a screwless DE-9, similar to the type of port actually found on the
Sega Genesis console.

The toolbar now shows 48x48 icons. This helps to reduce the amount of
wasted space when lots of Team Player devices are connected.

The "none" icon is now available in both 48x48 and 64x64. I rearranged
the gens/ icons directory a bit to use a directory hierarchy by size.