[libgens] IO/Io*.?pp: Marked the read data/ctrl functions and zomgSaveMD() as const.
IoMegaMouse::readData(): Moved the latchMouseMovement() call to
IoMegaMouse::writeData(). This is required in order to mark readData()
as constant, and it's more accurate, since the mouse doesn't know when
the M68K reads the data lines on the controller port.

Also, it's more resilient in the case of games that incorrectly read
the data port more than once without incrementing the mouse's counter,
since it will only latch when the counter hits 1 instead of latching
every time the data port is read when the counter equals 1.