[gens-qt4] Config/CtrlConfig.cpp: Added defaults for Port 1 and Port 2.
Port 1 defaults to a 6-button gamepad.
Port 2 defaults to a 3-button gamepad.

TODO: Come up with better default keymappings.

CtrlConfigPrivate::load(): If an entry isn't found (or the type is
invalid), a default controller type and keymap will be used.
The default types and keymaps are stored in
CtrlConfigPrivate::Def_CtrlTypes[] and CtrlConfigPrivate::Def_CtrlKeys[].