[gens-qt4] Config/RecentRoms.cpp: Initial "Recent ROMs" class.
This class manages the "Recent ROMs" function. Currently, it supports
updating an internal list of ROMs, and can load and save information
to/from a QSettings file.

EmuManager::openRom(const QString& filename): Update the Recent ROMs
list when a ROM is opened.

TODO: Don't do this if the ROM couldn't be loaded.