[libgens] EmuMD::setRegion_int(): Set/clear the VDP video mode bit based on video mode.
The VDP status register's video mode bit was always set to 0, which was
equivalent to NTSC. Most games check the system version register, but a
few games check the VDP status register for region lockout. Specifically,
"Adventures of Mighty Max, The (Europe).md" (CRC32: 24f1a3bb) checked the
VDP status register, which prevented the game from working.

This commit sets the video mode bit in the VDP status register to 0 for NTSC
and 1 for PAL.

This is a regression from Gens/GS, which set this bit correctly.

This bug was reported by andlabs in #techies.