[libgens] cpu/M68K_Mem.cpp: Virtualize all the M68K memory banks.
This is similar to the old function pointer system, but instead of
function pointers, bank type IDs are used. This allows for the same
optimizations as the improved system while providing the flexibility
of the function pointer system.

TODO: Possibly change the bank size from 512 KB to 128 KB.
"16 Ton", a SegaNet game, reportedly has 128 KB RAM in the cartridge
in addition to the 256 KB ROM, so in order to emulate it properly,
we'll need to use a smaller bank size.