[gens-qt4] Added an option to disable SRAM/EEPROM.
The option is located in the Options menu.

Note that even if SRAM/EEPROM is disabled, if an SRAM or EEPROM
file for the game is found, it will be loaded on startup. The game
simply won't be able to see its contents unless SRAM/EEPROM is
re-enabled later on.