[libgens] VdpRend_m5::T_Make_Sprite_Struct(): Mask the high bit of the sprite link field.
This fixes the main game in "Dino Land". For some reason, "Dino Land" sets
the high bit of some sprites. The MD VDP ignores the high bit, but we weren't
ignoring it, resulting in the sprite loop terminating early. (High bit would
force the link number to be >= 128, which is greater than max_spr.)

TODO: Handle wraparound if Spr_Addr is >= 0xFC00. (reg5 >= 0xFE)