[libgens] VdpPalette: Added a new property for the MD color mask, mdColorMask.
Externally, this property is set using a boolean: false for full color range,
and true for LSB only. Internally, the property is managed as a uint16_t for
optimization, since this uint16_t is AND'd with the actual MD colors before
saving the color in the internal MD palette.

This change eliminates most of the dependencies on the Vdp class from VdpPalette.

TODO: Eliminate the MarkCRamFlag() calls. Add a new external dirty flag or
something similar.