[libgens] VdpPalette::T_recalcFull_32X(): Updated to work with VdpPalette.
This function was originally T_Recalculate_Palette_32X().

NOTE: MD color scaling was removed from the Gens/GS II version. It makes
less of a difference with the 32X's 15-bit color than it does with the MD's
9-bit color. The 32X color is now always scaled using a simple algorithm
that uses the 32X 5-bit color component as the top 5 bits of the RGB 8-bit
color component, and then duplicates the high 3 bits and uses that as the
low 3 bits of the RGB 8-bit color component.

TODO: This function is untested!
TODO: Port T_Adjust_CRam_32X(), and rename it to T_update_32X().