[libgens] Vdp/VdpPalette.cpp: New functions T_recalcFull_TMS9918() and T_update_TMS9918().
These functions use the hard-coded analog color palette. They will be
used if VdpPalette's palette mode is set to PALMODE_TMS9918.

NOTE: If VdpPalette's palette mode is set to PALMODE_SMS and a TMS9918
mode is set in the Vdp, the Vdp should call VdpPalette::initSegaTMSPalette()
to load the default SMS TMS9918 colors. These colors are based on
6-bit RGB, which is different from the original TMS9918 color palette,
which was based on analog circuitry.

TODO: Test these functions!