[libgens] Vdp/VdpPalette.cpp: Copy the TMS9918 palette to all four MD palette lines.
This is useful for debugging, and it won't cause much overhead because
the TMS9918 palette can't be updated at runtime.

NOTE: If CRam is written, m_dirty.active will be set to true and the
palette will be updated again. This shouldn't happen when emuiating
consoles with TMS9918 VDPs, since the TMS9918 doesn't have CRam.