[glew] Updated the internal copy of GLEW to version 1.7.0.
There's a bug in glew-1.6.0 and glew-1.7.0's custom code generation
function that essentially breaks it. The bug was caused by switching
to perl instead of sed, but not changing the regular expressions to
Perl-compatible. I manually reverted the Makefile to use sed, so I
was able to generate the customized version of GLEW.

A patch has been submitted, so it should be fixed for the next release.