[libgens] mdZ80: Converted IFF into a single uint8_t.
IFF now uses the same format as ZOMG's IFF: [0 0 0 0 0 0 IFF2 IFF1]

I finally figured out why Gens used FLAG_P for the IFF1 and IFF2 values:
IFF2 is copied to the P/V flag when either the LD A,I or LD A,R instructions
are executed. It now uses bit manipulation, which might be slightly slower,
but these instructions aren't too common, so it shouldn't be a problem.

mdZ80.c::mdZ80_reset(): Use offsetof() to determine how much of the
Z80 context we should erase. This will ensure that things won't break
if the Z80_CONTEXT struct is resized again, and it probably will be.
(For example, the registers will be shrunken from 32-bit to 16-bit.)