[libgens] mdZ80: Z80_CONTEXT (now mdZ80_context) is now an opaque struct.
Hiding the implementation details makes it easier to handle layout changes
due to architecture-specific optimizations (such as the AF layout for
lahf/sahf optimization on x86). The only downside is that accessing the
Z80 registers from other parts of the emulator is now slightly slower.

Link-time optimization should reduce the impact of that slowdown.

mdZ80_init() has been renamed to mdZ80_new(), which returns a newly-allocated
mdZ80_context*. A corresponding mdZ80_free() function has been added.

Added functions to get and set the mdZ80 context registers.

Also, fixed an error in mdZ80_init() when initializing Fetch[] that ended
up writing 4 times the number of entries in Fetch[]. This didn't cause
any problems when the Z80_CONTEXT was statically allocated, but with the
new dynamic allocation, it resulted in heap corruption.