[libgens/tests] test_Vdp_SpriteMaskingTestRom.cpp: Initial Sprite Masking & Overflow Test ROM test suite.
This test suite loads the VRAM, CRAM, and registers from Nemesis'
Sprite Masking & Overflow Test ROM. It will use screen-scraping
to determine if the tests passed or failed.

The VRAM is compressed using gzip and stored in test_Vdp_SpriteMaskingTestRom_data.c,
so zlib is required for this test suite.