[libgens/tests] Test_SpriteMaskTestRom::runTestSection(): Reset the VDP line counters before rendering each frame.
Previously, the line counters were only reset when the test was started,
which means it worked fine for the first test (H32 with sprite limits),
but the second test (H32 with no sprite limits) showed the same results,
and the H40 tests were all wrong (all pixels blue).

This ensures that the VDP frame is actually re-rendered.