[mdZ80] mdZ80.c: Split hard-reset and soft-reset handling into separate functions.
The Z80 doesn't clear all registers when the !RESET line is asserted.
I'm not positive about certain registers like the regular data registers
and the shadow registers, but we have documentation for AF, PC, and SP.

I'll write an MD test program later to check the values of these registers
on !RESET so someone can test the actual hardware.

Updated libgens and gens-qt4 to handle the Z80 reset split.