[gens-qt4] config/ConfigStore.cpp: Fixed Qt slot invocation.
SLOT() includes a "1" prefix indicating the method is a slot, plus the
parameters attached as a suffix. QMetaObject::invokeMethod() doesn't
like this, and prefers to build its own function signature.

I added a new function, InvokeQtMethod(), which is partially based on
QMetaObject::invokeMethod(), except it doesn't do all the fancy signature
building. It uses the method name as-is, minus the signal/slot prefix.

TODO: Verify that Qt methods are valid in registerChangeNotification()?

Renamed SignalMap.obj to SignalMap.object, and SignalMap.slot to SignalMap.method.