[gens-qt4] EmuManager::loadRom_int(): Initialize the VDP graphics settings.
This needs to be done when creating an emulation context now, since the
Vdp class is no longer static.

Graphics settings are finally applied on application startup, instead of
assuming defaults until one of the settings is changed. This hasn't worked
properly since the Vdp class was converted from static to non-static.