[gens-qt4] GeneralConfigWindow: Use an icon toolbar on all platforms.
The icon toolbar looks better than the tab layout. Also, this cuts
down on OS-specific code, and should result in fewer issues on the
Mac version, since it doesn't have to delete the tab widget and
create the stacked widget on the fly.

TODO: Test on Mac OS X to make sure this didn't break anything.