[gens-qt4] Config/ConfigStore.cpp: Implemented paramter validation.
The Validate() function returns either the validated value or an invalid
QVariant, which is then either discarded by the set() function, or replaced
with the default value by the load() function.

Copied EmuManager::IsRegionCodeOrderValid() to ConfigStorePrivate.
The EmuManager version will be deleted in the next commit, along with
the rest of the validation code outside of ConfigStore.