[libgens] Vdp::T_Get_Y_Offset(): The VScroll bug code does work for Oerg's MDEM 2011 demo.
Turns out the problem was that the final ROM has sprites to mask the broken
column on the left. I needed an earlier build that didn't have these sprites
in order to test the bug.

It now matches the original MD1/MD2 hardware with the VScroll bug enabled.

Video demonstrating the VScroll bug:
- http://www.youtube.com/watch?v=UsW8i7zsY8w