[gens-qt4] ConfigStore::Validate(): Use QString::toInt() and QString::toUInt().
This allows Validate() to parse hexadecimal (or octal) values.

This fixes a problem where changing any setting other than region code
in GeneralConfigWindow while the region code was set to auto-detect
resulted in the emulator printing the auto-detect region code message.
It turns out that the real reason this happened is because the auto-detect
order was loaded as 0, since it didn't pass validation when saved as a
hexadecimal string.

This is a regression from commit 01a8c51a8415c979aa129a28fe90cb0310026fbe.
([gens-qt4] Config/ConfigStore.cpp: Implemented paramter validation.)