[gens-qt4] RecentRomsMenuPrivate::update(): Don't insert a '\t' character after the system ID.
This works fine when using the default Qt theme on Linux and Windows,
but when using KDE 4.x's Oxygen theme, the ROM name is right-aligned.
This is probably because the '\t' tells the theme that the remainder
of the QAction text is the shortcut key, and Oxygen right-aligns the
shortcut key display.

Without the '\t', different system IDs will have different starting
positions for the ROM name, but that's better than right-alignment.