[libgens] Vdp/: Implement VDP_Reg.m5.Set1, bit 5 (LMSK).
This bit is used in Mode 4 to mask the left-most column. It's used in many
Master System games due to a bug in the horizontal scrolling implementation.
Specifically, horizontal scrolling does not wrap around properly if the
HScroll value is not a multiple of 8. For instance, if HScroll is 4, then
the first four pixel columns will not be the second half of the last column
of tiles; instead, it will be the background color. The LMSK bit works
around this by hiding the broken column entirely.

It turns out that this functionality is still present in Mode 5, even though
horizontal scrolling was fixed. Hence, we should implement LMSK, since it
will help to weed out broken hacks and homebrew.

This fix was suggested by Jorge.