[gens-qt4] Config/ConfigStore.cpp: Allow certain settings to be reset to their current value.
Added a new field DefaultSetting::allow_same_value, which controls this
behavior. If set, then oldValue checking is skipped, and signals are
emitted regardless of whether or not the setting has actually changed.

This is useful for e.g. "Savestate/saveSlot", since the user should
be able to press the key corresponding to the current save slot in
order to see the preview image for that savestate.

"Savestate/saveSlot" is currently the only setting with the allow_same_value
field set to true.