[libgens] Util/MdFb.hpp: Updated to use reference counting.
Added a clear() function. Vdp::rend_reset() was using memset() to clear
MD_Screen, which was wrong anyway because it's a C++ class, but with
the conversion to pointers, it ended up clearing the pointer instead
of the framebuffer data.

Updated all of libgens and gens-qt4 for the reference counter changes.