[gens-qt4] Config/ConfigStore.cpp: Added a load-only setting feature.
If no_save is set, the setting will not be saved.

This will be used for "iKnowWhatImDoingAndWillVoidTheWarranty", which
enables advanced low-level settings at the cost of voiding the warranty.

(Then again, since this *is* GPL-licensed software, there is no warranty
to begin with. It's more of a technical support thing, and also kind of
an in-joke.)