[gens-qt4] GensQApplication_win32.cpp: Enable heap metadata protection.
I haven't tested this on a Windows machine, but it should work.
If it results in crashing, it means that there's a bug somewhere
in the code and that should be fixed.

This feature requires Windows XP SP3, Windows Vista, or later.

Reference: http://msdn.microsoft.com/en-us/library/bb430720.aspx