[gens-qt4] gqt4_main.cpp::gens_main(): Set Qt::AA_X11InitThreads.
On Qt 4.8, Qt::AA_X11InitThreads tells Qt to initialize X11 threads itself.
This is needed to prevent threading issues with e.g. QPixmap.

If Gens/GS II is compiled with an older version of Qt, X11 threads will be
initialized manually with XInitThreads(), but the attribute will be set
anyway, since the user might upgrade the Qt libraries later on without
recompiling Gens/GS II.

This is partially based on a recent commit to LibreOffice:
http://cgit.freedesktop.org/libreoffice/core/commit/?id=6163c62a52794725188faf82ee12852f4b2cd023
(let Qt call XInitThreads(), so that it knows it's been called (fdo#40298))