Revert [libgens] EmuMD::EmuMD(): Floor the cycles-per-line values instead of rounding them.
This reverts commit 538a5026bc9ae078a56b50f85a53267c6c50d015.

It seems that using floor() causes distortion on the title screen of
"Beavis and Butt-head" (U).

Also, Jorge says that CPL is 3420 master clock cycles regardless of
NTSC/PAL, so perhaps we should simply use that value instead of
calculating CPL based on region. (Right now it gets 3416 CPL on NTSC
and 3410 CPL on PAL.)