CtrlConfigWindow: Implemented reload().
This function isn't currently assigned to any button, but it's
now being used by the CtrlConfigWindow constructor to load the
initial port settings.

This function also checks to make sure that if a Teamplayer or 4-Way Play
port was selected, and the appropriate device was removed due to the reload,
an appropriate base port is selected instead.