IoManager: Converted uint8_t devBtnCount[] to IoDevInfo ioDevInfo[].
ioDevInfo[] contains a FourCC that refers to the controller type,
as well as button count and whether or not the controller is fully
'usable' in this build.

If the controller is not usable, it will be disabled in release builds.

The FourCC will be used in the configuration file in order to decouple
internal controller IDs from the configuration. It will also be used
in the ZOMG save file format, once I implement controller data saving.