[libgenscd] CdDrive: Implemented an OS-specific function hasDiscChanged().
Linux provides CDROM_MEDIA_CHANGED, which usually works.

Both Windows and LInux *should* work properly with the
MMC GET_EVENT_STATUS_NOTIFICATION command, but this
command seems to be broken.

NOTE: The GUI should call CdDrive::forceCacheUpdate() if it detects
a disc change through e.g. UDisks or WM_DEVICECHANGE. This forces
the cache to be updated regardless of the result of hasDiscChanged().