[libzomg] Initial ZOMG byteorder rework.
Functions that save 16-bit or 32-bit data should have an explicit byteorder
passed to them in order to ensure the data is swapped correctly, instead of
relying on a boolean variable that may or may not be accurate.

ZomgByteorder_t supports 8, 16LE, 16BE, 32LE, and 32BE, plus has aliases
16H and 32H for host-endian.

ZomgLoad: Updated all 16-bit load functions to use ZomgByteorder_t.
A templated function LoadMemByteswap<>() is used to ensure the data
is byteswapped correctly.

[libgens] GensZomg: Updated to work with the ZomgLoad changes.