[libzomg] zomg_vdp.h::Zomg_VdpCtrl_16_t: Added a "dma_access" field.
LibGens::Vdp::VDP_Ctrl.Access is 16-bit. The high byte is exclusively
for DMA, while the low byte is for standard operations. The high byte
should thus be stored separately. (The ordering of the VdpCtrl struct
is such that it's mostly compatible with Sega 8-bit systems, though
it obviously isn't directly compatible due to other differences.)

Currently, the high byte isn't stored at all, which might be a bit
problematic later on.