[libgens] SMS/MD VDP starts at visible line 0.
In the initial VDP rewrite, I changed it so the VDP would start at
display line 0, and visible line 0 would be drawn a few lines later.
This matched the NES, but was actually wrong for the SMS/MD.

The only notable issue caused by this was that if Team Player adapters
were connected and the Multitap I/O test ROM was loaded, some of the
text would be written to CRAM instead of the display, resulting in
weird color issues. With this fix, the test ROM doesn't show this
issue anymore.

VDP_Lines_t:
- Removed all of the "Display" line variables except for totalDisplayLines.
- Added border variables to make border calculation a bit easier.