Added an option to update the palette in VBlank only.
This simulates the behavior of the Genecyst emulator, which e.g.
only rendered underwater levels using the underwater palettte
if the entire screen was submerged.

The option is in the "warranty is void" section, so it will only
show up in GeneralConfigWindow if the user knows what they're
doing and has voided the warranty.