Backported many CMake changes from GCN MemCard Recover.
Includes lots of CFLAGS changes as well as some PDB improvements.
In particular, -Zi is used during compilation so PDBs are used
instead of including the debug symbols in the objects themselves.

gtest: Set gtest_force_shared_crt to force gtest to link to the
dynamic CRT on MSVC instead of the static CRT, since it usually
links to the static CRT when compiled as a static library.

TODO: CMake uses a dumb default filename for PDBs.
I'm using some ad-hoc code to fix this right now, but it needs
to be abstracted into a macro.