[libgens] Vdp::Write_Ctrl(): Register writes affect the 'code' and 'address' registers.
Don't set them to 0; instead, move the existing latched control word
code to before the setReg() code, then call setReg() if the control
word is a register write.

This doesn't fix any tests in VDPFIFOTesting, but it does eliminate
some odd garbage in the background after going to the second page.