[libgens] Initialize TMSS before resetting the 68000.
Starscream initializes the internal program counter on reset, so we
have to initialize TMSS first; otherwise, if the TMSS enable/disable
option is changed, and the user does a Hard Reset, the first reset
will show a black screen. If the user does another reset (either hard
or soft), it will reset correctly.

Initializing TMSS before the CPU reset works around the problem.