[gens-qt4] GeneralConfigWindow_p.cpp: Minor improvements to ROM file status.
If the filename is empty, exit early instead of trying to check if the
file exists.

Use QFileInfo to get information about the file.

If the file is a directory, say that instead of trying to identify the
ROM, which won't work. It also showed an incorrect filesize of
2,147,483,647 bytes (on Linux).

TODO: Check for device files and skip them.