[libgens] Vdp: Use the Sprite Attribute Table cache instead of our custom cache.
The cache is updated prior to the line being rendered. This is how it
works on the actual system, and it will definitely help when I add
support for TMS9918A and Sega Master System modes later.

Removed Sprite_Visible[]. The sprites on each line are calculated and
stored in an array of SprLineCache_t, which used to be Sprite_Struct.

sprLineCache[]: Two arrays of 20 SprLineCache_t objects. The VDP can
render up to 20 sprites in H40 (16 in H32). The first dimension is the
LSB of the line being cached, since the VDP renders the line that was
previously cached while caching the next line.

sprCountCache[]: Indicates how many sprites are visible in the associated
sprLineCache[] line.

T_Update_Sprite_Line_Cache<>(): Replaces both T_Make_Sprite_Struct<>()
and T_Update_Mask_Sprite<>(). The former obtained the sprite data for
all sprites onscreen, and the latter determined which sprites were
being displayed on the current line. This is no longer needed with
the line cache.

TODO:
- SpriteMaskingTestRom fails on tests 2, 3, and 6, all on the second
  column. ("sprites that shouldn't be shown")
- Related: Sonic's on top of the logo on Sonic 1's title screen.
- Allow the user to disable the sprite limit. This will require
  extending sprLineCache[] to 80 elements.