[libgens] Vdp::dbg_writeVRam_16(): Update the Sprite Attribute Table.
This fixes most of the tests in VdpSpriteMaskingTest; however, two tests
are failing with sprite limits turned *off*:
- Test #1: Max Sprites per Line (H32/H40)
- Test #9: Max Sprites per Frame (H32)

These are failing because we're only caching a maximum of 20 sprites per
line, even with sprite limits disabled. The sprite line cache has to be
extended to 80 sprites in order to match the previous behavior with
sprite limits disabled.