[libgens] VdpPrivate::T_Update_Sprite_Line_Cache: Extend sprite limits if the user disabled them.
This fixes the following tests with sprite limits turned off:
- Test 1: Max Sprites per Line (H32)
- Test 9: Max Sprites per Frame (H32)

Test 1 in H40 is still broken, since that requires extending sprLineCache.

NOTE: While Test 9 (H32) passes in VdpSpriteMaskingTest, it does NOT
go over the frame limit with sprite limits turned off in the actual
emulator. This seems to be an issue with the Sprite Attribute Table.