[libgens] Vdp: Don't allow frame sprite limits to be disabled.
Due to the Sprite Attribute Table mask, sprite limits can't really be
disabled, since sprites 64-79 can't be updated in the SAT when the VDP
is set to H32. If the sprites were updated during H40 and then H32 is
set, they'll remain there, but those sprites still won't show up.

[libgens/tests] VdpSpriteMaskingTest: Update test 9 such that the expected
results is always "pass" regardless of the sprite limit setting.