[libgens] EEPRom: Initial rewrite of the I2C EEPROM code.
The current code is an unmaintanable mess and has some issues.
For example, NBA Jam TE (Rev00) creates a 2048-byte EEPROM file,
even though it has a 24C04 (4 Kbit, 512 bytes) EEPROM.

The new code will be based on the original datasheets, and changes
will be applied on proper line changes (e.g. /SCL transitions),
not on M68K read/write cycles. The EEPROM has no way to tell when
an M68K read/write cycle ends, so it's all dependent on /SCL.

Currently, the class has been restructured to have a private class,
and the actual read/write code is gone. I'll start with the X24C01,
which is used by "Mega Man: The Wily Wars".