[libgens/tests] EEPRomI2CTest: Do NOT acknowledge the received data if we want to go into START/STOP afterwards.
The X24C01 and 24C01A/02/03/04/16 datasheets both say that the last byte
in the read should *not* be acknowledged. This is because the EEPROM will
start transmitting the next byte, and if the next byte starts with a 0 bit,
that will prevent a START or STOP condition from being recognized.