[libgens] EEPRomI2CPrivate::processI2Cbit(): Send ACK when /SCL=0.
"NBA Jam" checks for ACK when /SCL=0. The datasheet says ACK shouldn't
be checked until /SCL=1, but clearly we need to set it when /SCL=0
because otherwise NBA Jam locks up in an infinite loop trying to
read from the EEPROM.

This fixes "NBA Jam". The generated EEPROM file appears to be correct.