[libgens] EEPRomI2CTest: Added Random Read tests for 24C01.
FIXME: The tests are failing due to an off-by-one in the output
bit shifting. It seems that only 7 of the 8 bits are being shifted
out for some reason. For example, in randomReadEmpty, all of the
empty bytes (0xFF) are being read as 0x7F, and in randomReadFull,
the off-by-one is causing all sorts of randomness.