[libgens] EEPRomI2CPrivate::processI2CShiftIn(): Fixed dev_addr masking for MODE2.
Make sure we only use the relevant three bits from the command word,
and invert the size mask in order to use it to mask the device address.

This fixes the 24C04-24C16 Random Read tests, and it also fixes a minor
issue where "NBA Jam TE" (REV00) was only writing 256 bytes in the
EEPROM. It's now writing the full 512 bytes. (24C04)