[libgens] Vdp: Fixed sprite rendering in interlaced mode.
We *do* have to use the line number adjusted for IM2 in order to
determine which sprites to draw, since the Y position is extended
by one bit in IM2.

FIXME: If T_Update_Line_Sprite_Cache<>() is called with -1, we should
adjust it to line 0 or 1 depending on odd/even frames.