[libgens] Vdp: Added a secret option "enableInterlacedMode".
If disabled, games that use interlaced mode (e.g. Sonic 2's 2P VS)
will show garbage on the screen, similar to Genecyst.

Disabling this option does NOT prevent the odd/even frame bit
or the IM2 VCounter from working properly. It only affects
the way the image is rendered. (...and it might affect sprite
collisions and/or overflow, but that's highly unlikely.)