[libgens] TmssReg: Moved the TMSS loading code here from EmuMD.
EmuMD's initTmss() is still needed in order to update the version
register and to call M68K_Mem::UpdateTmssMapping().

TmssReg supports loading up to 512 KB for the TMSS. The allocated buffer
is always a power of two, with a minimum of 4 KB. Extra space is filled
with 0xFF bytes.

Added readByte() and readWord() functions to read the TMSS ROM.