[gens-qt4] GeneralConfigWindow: Improved the TMSS detection algorithm.
If the TMSS ROM is a multiple of 2 KB and less than 512 KB, it checks
the first 2 KB to make sure it's correct. Then, it checks consecutive
2 KB blocks:
- Odd blocks: Must be 0x00, 0xFF or the TMSS ROM.
- Even blocks: Must be 0x00 or 0xFF.

This detects certain overdumps. tristanseifert reports that when he
dumped the TMSS ROM on his Genesis, the first 2 KB was TMSS data,
the second 2 KB was open bus, and then it repeated. The code he used
actually dumped 2 KB data, 14 KB 0x00, but that's detected properly
as well.