[gens-qt4] GensCtrlCfgWidget: Fix up logical vs. physical button mapping.
When loading the key, we have to load it from the physical index
to the logical index. When saving the key, we have to save it from
the logical index to the physical index.

This fixes up all the previous issues with controller configuration.

TODO: With this layout, SMS Pause/Start is mapped to the same key as
the "A" button on 3BTN/6BTN.